7 scaffold the note page (#21)
* feat: Adding tiptap to the project

Tiptap is a rich text editor

* feat: Adding a Rich Text Editor component

Adding the rich text editor component and giving a minute amount of styling to distinguish it when it is active on the page.

* feat: Renaming the notebookId to id

This simplifies the url

* housrkeeping: Restructuring project

Giving related independent components such as menu-item their own folder within the menu folder.
Renaming index files to the component name even though it's repetition.

* housekeeping: Updating imports

* refactor: Extract NoteDataType

As it's needed in more than one place now

* housekeeping: Refactoring component names

Plus of course updating imports

* feat: Creating a dispatcher type

Allows me to pass setState as props without typescript grumbling

* feat: Updating Editor for presentation

Adding dummy function for sending data to ensure collecting the updated note correctly.

* fix: The issue with the note ID

Ho, ho, ho, what fun this was. It turns out that passing something as a number into params and then suggesting that the type of the id is a number doesn't mean it's a number. The type of any param is string and you'd better remember that!

* Rename header.tsx to Header.tsx